use mutex for downloading profile photos
This prevents an issue where some concurrent asset downloads were causing the server to disconnect us.

The recommendation of the developer seems to only work for downloadMedia, but I made the change there as well since we are using WS.

https://github.com/gram-js/gramjs/issues/227#issuecomment-1004278484

Performance didn't seem heavily affected, and only on first load. This can be verified by uncommenting line #364 for api.js